pyproject.toml
[project]
name = "otel-fastapi-link-middleware"
version = "0.1.5"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Ranjith BR", email = "ranjith19@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["opentelemetry-sdk>=1.32.1", "starlette>=0.40.0"]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
